+++ src/main/java/com/example/solidconnection/auth/service/AuthService.java
@@ -2,6 +2,7 @@ package com.example.solidconnection.auth.service;
 
 import com.example.solidconnection.auth.dto.ReissueRequest;
 import com.example.solidconnection.auth.dto.ReissueResponse;
+import com.example.solidconnection.auth.token.TokenBlackListService;
 import com.example.solidconnection.common.exception.CustomException;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import lombok.RequiredArgsConstructor;
