+++ src/main/java/com/example/solidconnection/auth/token/TokenBlackListService.java
@@ -1,5 +1,6 @@
-package com.example.solidconnection.auth.service;
+package com.example.solidconnection.auth.token;
 
+import com.example.solidconnection.auth.service.AccessToken;
 import com.example.solidconnection.security.filter.BlacklistChecker;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
