+++ src/test/java/com/example/solidconnection/security/filter/SignOutCheckFilterTest.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.security.filter;
 
 import com.example.solidconnection.common.exception.CustomException;
-import com.example.solidconnection.security.config.JwtProperties;
+import com.example.solidconnection.auth.token.config.JwtProperties;
 import com.example.solidconnection.support.TestContainerSpringBootTest;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
